services/chat-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# Обработка файлов
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1

# OCR для PDF
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Экспорт документов
reportlab==4.0.7
python-docx==1.1.0

# Дополнительные утилиты
aiofiles==23.2.1
python-magic==0.4.27
aiohttp==3.9.1

# Для умной токенизации
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2
